+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/basic/tile/TileCampfire.java
@@ -8,7 +8,10 @@ import com.codetaylor.mc.athenaeum.network.tile.data.TileDataItemStackHandler;
 import com.codetaylor.mc.athenaeum.network.tile.spi.ITileData;
 import com.codetaylor.mc.athenaeum.network.tile.spi.ITileDataItemStackHandler;
 import com.codetaylor.mc.athenaeum.network.tile.spi.TileDataBase;
-import com.codetaylor.mc.athenaeum.util.*;
+import com.codetaylor.mc.athenaeum.util.BlockHelper;
+import com.codetaylor.mc.athenaeum.util.SoundHelper;
+import com.codetaylor.mc.athenaeum.util.StackHelper;
+import com.codetaylor.mc.athenaeum.util.TickCounter;
 import com.codetaylor.mc.pyrotech.interaction.api.InteractionBounds;
 import com.codetaylor.mc.pyrotech.interaction.api.Transform;
 import com.codetaylor.mc.pyrotech.interaction.spi.*;
@@ -724,7 +727,7 @@ public class TileCampfire
 
     private boolean doItemStackValidation(ItemStack itemStack) {
 
-      return OreDictHelper.contains("logWood", itemStack);
+      return ModuleTechBasicConfig.CAMPFIRE.isValidFuel(itemStack);
     }
 
     @Override
@@ -911,7 +914,7 @@ public class TileCampfire
     @Override
     public ItemStack insertItem(int slot, @Nonnull ItemStack stack, boolean simulate) {
 
-      if (!OreDictHelper.contains("logWood", stack)) {
+      if (!ModuleTechBasicConfig.CAMPFIRE.isValidFuel(stack)) {
         return stack;
       }